+++ inst/examples/gg.sp.R
@@ -29,7 +29,7 @@
 
     # Load pantropical dolphin data
 
-    mexdolphin <- inlabru::mexdolphin
+    mexdolphin <- inlabru::mexdolphin_sp()
 
     # Plot the pantropical survey boundary, ship transects and dolphin sightings