+++ tests/testthat/test-lgcp_1d_nonlinear.R
@@ -1,7 +1,7 @@
 test_that("Mexdolphin: Hazard rate detection function", {
   skip_on_cran()
   local_bru_safe_inla()
-  data(mexdolphin, package = "inlabru", envir = environment())
+  mexdolphin <- inlabru::mexdolphin_sp()
 
   sig <- function(x) bru_forward_transformation(qexp, x, rate = 1 / 8)
   hr <- function(distance, sigma) {